Refactor tabular data
